pyproject.toml
[project]
name = "manageprojects"
dynamic = ["version"]
description = "Manage Python / Django projects"
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
authors = [
    {name = 'Jens Diemer', email = 'mamageprojects@jensdiemer.de'}
]
requires-python = ">=3.11"
dependencies = [
    "cookiecutter>=2.4.0",  # https://github.com/cookiecutter/cookiecutter
    "tomlkit",
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py

    # https://github.com/akaihola/darker
    # https://github.com/ikamensh/flynt
    # https://github.com/pycqa/isort
    # https://github.com/pygments/pygments
    "darker[flynt, isort, color]",

    "autoflake",  # https://github.com/PyCQA/autoflake
    "pyflakes",  # https://github.com/PyCQA/pyflakes
    "autopep8",  # https://github.com/hhatto/autopep8
    "flake8",  # https://github.com/pycqa/flake8
    "flake8-bugbear",  # https://github.com/PyCQA/flake8-bugbear
    "pyupgrade",  # https://github.com/asottile/pyupgrade
    "refurb",  # https://github.com/dosisod/refurb
    "codespell",  # https://github.com/codespell-project/codespell
    "mypy",  # https://github.com/python/mypy

    "cli-base-utilities",  # https://github.com/jedie/cli-base-utilities
    "tyro",  # https://github.com/brentyi/tyro
    "rich",  # https://github.com/Textualize/rich
]

[dependency-groups]
dev = [
    "python-gnupg",  # https://github.com/vsajip/python-gnupg
    "bx_py_utils",   # https://github.com/boxine/bx_py_utils
    "urllib3", # for deny_requests.deny_any_real_request() in tests
    "uv",  # https://github.com/astral-sh/uv
    "setuptools",  # https://github.com/pypa/setuptools
    "build",  # https://github.com/pypa/build
    "tox",  # https://github.com/tox-dev/tox
    "tox-uv",  # https://github.com/tox-dev/tox-uv
    "coverage",  # https://github.com/nedbat/coveragepy
    "autopep8",  # https://github.com/hhatto/autopep8
    "pyupgrade",  # https://github.com/asottile/pyupgrade
    "flake8",  # https://github.com/pycqa/flake8
    "flake8-bugbear",  # https://github.com/PyCQA/flake8-bugbear
    "pyflakes",  # https://github.com/PyCQA/pyflakes
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "pip-audit",  # https://github.com/pypa/pip-audit
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine
    "pre-commit",  # https://github.com/pre-commit/pre-commit
    "typeguard",  # https://github.com/agronholm/typeguard/

    # https://github.com/akaihola/darker
    # https://github.com/ikamensh/flynt
    # https://github.com/pycqa/isort
    # https://github.com/pygments/pygments
    "darker[flynt, isort, color]",

    # Work-a-round for: https://github.com/jazzband/pip-tools/issues/1866
    # see also: https://github.com/jazzband/pip-tools/issues/994#issuecomment-1321226661
    # backports.tarfile is needed for python <3.12
    'backports.tarfile', # via jaraco-context -> keyring -> twine
]

[project.urls]
Documentation = "https://github.com/jedie/manageprojects"
Source = "https://github.com/jedie/manageprojects"

[project.scripts]
manageprojects_app = "manageprojects.__main__:main"
manageprojects_dev = "manageprojects.cli_dev:main"

[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=7.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["manageprojects*"]

[tool.setuptools.dynamic]
version = {attr = "manageprojects.__version__"}


[tool.cli_base.pip_audit]
# https://github.com/jedie/cli-base-utilities/blob/main/docs/pip_audit.md
requirements=["requirements.dev.txt"]
strict=true
require_hashes=true
ignore-vuln=[]


[tool.cli_base]
version_module_name = "manageprojects"  # Used by "update-readme-history" git hook


[tool.bx_py_utils.doc_write]
docstring_prefix = 'DocWrite:'
macro_prefix = 'DocWriteMacro:'
output_base_path = './docs/'
search_paths = ['./manageprojects/']
delete_obsolete_files = true


[tool.darker]
src = ['.']
revision = "origin/main..."
line_length = 119
color = true
skip_string_normalization = true
diff = false
check = false
stdout = false
isort = true
lint = [
    "flake8",
]
log_level = "INFO"


[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files/#pyprojecttoml-preferred-format
atomic=true
profile='black'
skip_glob=['.*', '*/htmlcov/*']
known_first_party=['manageprojects']
line_length=119
lines_after_imports=2


[tool.coverage.run]
branch = true
parallel = true
concurrency = ["multiprocessing"]
source = ['.']
command_line = '-m unittest --verbose --locals --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.tox]  # https://tox.wiki/en/latest/config.html#pyproject-toml
isolated_build = true
env_list = ["3.13", "3.12", "3.11"]
skip_missing_interpreters = true

[tool.tox.env_run_base]
runner = "uv-venv-lock-runner"
with_dev = true
commands = [
    ["python3", "-m", "coverage", "run", "--context", "'{envname}'"]
]


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests']


[manageprojects] # https://github.com/jedie/manageprojects
initial_revision = "8438598"
initial_date = 2022-11-20T20:01:37+01:00
cookiecutter_template = "https://github.com/jedie/cookiecutter_templates"
cookiecutter_directory = "uv-python"
applied_migrations = [
    "cf44769", # 2023-01-11T08:15:06+01:00
    "cd164cd", # 2023-02-18T11:55:57+01:00
    "d970456", # 2023-02-20T16:19:35+01:00
    "80d0a0c", # 2023-03-06T17:52:14+01:00
    "04d5a25", # 2023-03-07T16:25:36+01:00
    "be96b03", # 2023-03-17T08:17:11+01:00
    "588caaa", # 2023-04-04T15:40:44+02:00
    "00fbc5d", # 2023-04-07T12:35:09+02:00
    "39b7eab", # 2023-04-10T18:48:00+02:00
    "034be26", # 2023-08-05T21:37:12+02:00
    "8d0ebe1", # 2023-08-17T18:15:10+02:00
    "be3f649", # 2023-08-22T19:36:57+02:00
    "385f654", # 2023-10-08T21:09:24+02:00
    "d1ed4b1", # 2023-12-01T21:41:29+01:00
    "c538ae7", # 2023-12-21T20:21:58+01:00
    "7271d4c", # 2023-12-29T21:57:47+01:00
    "f8be3e0", # 2024-07-16T19:35:26+02:00
    "2cdc1d8", # 2024-08-25T19:00:41+02:00
    "56c3caa", # 2024-09-22T16:52:30+02:00
    "ff48b81", # 2024-11-09T19:08:01+01:00
    "690a4fb", # 2024-11-21T18:54:44+01:00
]

[manageprojects.cookiecutter_context.cookiecutter]
full_name = "Jens Diemer"
github_username = "jedie"
author_email = "mamageprojects@jensdiemer.de"
package_name = "manageprojects"
package_version = "0.3.2"
package_description = "Manage Python / Django projects"
package_url = "https://github.com/jedie/manageprojects"
issues_url = "https://github.com/jedie/manageprojects/issues"
license = "GPL-3.0-or-later"
_template = "https://github.com/jedie/cookiecutter_templates"
